requirements.txt
# ============================================================
# Artoo — Python Dependencies
# Python 3.12+  |  Last updated: March 2026
# Pinned to exact versions verified in production venv.
# ============================================================

# ── LangChain Core ───────────────────────────────────────────
langchain==1.2.10
langchain-core==1.2.14
langchain-aws==1.2.5
langchain-openai==1.1.10

# ── LangGraph ────────────────────────────────────────────────
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8

# ── MCP Integration ──────────────────────────────────────────
langchain-mcp-adapters==0.2.1
mcp==1.26.0

# ── AWS (Bedrock) ─────────────────────────────────────────────
boto3==1.42.54
botocore==1.42.54

# ── OpenAI ───────────────────────────────────────────────────
openai==2.21.0
tiktoken==0.12.0

# ── Pydantic ─────────────────────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.13.1

# ── Scheduler ────────────────────────────────────────────────
APScheduler==3.11.2

# ── Persistence ──────────────────────────────────────────────
SQLAlchemy==2.0.46

# ── HTTP / Async ─────────────────────────────────────────────
httpx==0.28.1
httpx-sse==0.4.3
aiohttp==3.13.3

# ── Logging ──────────────────────────────────────────────────
structlog==25.5.0

# ── Metrics API ──────────────────────────────────────────────
fastapi==0.131.0
uvicorn==0.41.0
python-multipart==0.0.22
sse-starlette==3.2.0

# ── Dashboard ────────────────────────────────────────────────
streamlit==1.55.0
pandas==2.3.3

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.2.1
tenacity==9.1.4
python-dateutil==2.9.0.post0
typing-extensions==4.15.0
requests==2.32.5

# ── Testing ──────────────────────────────────────────────────
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
moto[bedrock]==5.1.21
coverage==7.13.4
responses==0.26.0
